refactor: 모듈 구조 개선 및 이관작업
