+++ src/components/ScheduleTables/index.ts
@@ -1,0 +1,2 @@
+export * from "./ScheduleTable";
+export * from "./ScheduleTables";
