+++ src/components/SearchDialog/index.ts
@@ -1,0 +1,1 @@
+export * from "./SearchDialog";
