+++ src/data/index.ts
@@ -1,0 +1,1 @@
+export { default as dummyScheduleMap } from "./dummyScheduleMap";
