+++ src/main.tsx
@@ -1,7 +1,7 @@
 import { ChakraProvider } from "@chakra-ui/react";
 import { createRoot } from "react-dom/client";
 
-import App from "./App";
+import { App } from "./App";
 
 createRoot(document.getElementById("root")!).render(
   <ChakraProvider>
